fix bug; add relation events
